+++ views/elements/scaffold/data.html.php
@@ -1,4 +1,4 @@
-<table class="table table-striped table-condensed">
+<table class="table table-striped table-condensed table-hover table-selectable scaffold-data-view">
 	<colgroup>
 		<col width="170" />
 		<col width="*" />
@@ -11,7 +11,7 @@
 	</thead>
 	<tbody>
 	{{#each data}}
-		<tr data-key="{{@key}}" data-value="{{ this }}">
+		<tr data-id="{{@key}}" data-key="{{@key}}" data-value="{{ this }}">
 			<td class="key">{{@key}}</td>
 			<td class="value">{{ this }}</td>
 		</tr>