+++ views/elements/scaffold/view.html.php
@@ -1,1 +1,3 @@
-<?= $this->scaffold->render('data', array('data' => \lithium\util\Set::flatten($this->scaffold->object->data()))); ?>
+<?= $this->scaffold->render('meta', ['data' => $this->scaffold->object->data()]); ?>
+<hr />
+<?= $this->scaffold->render('data', ['data' => \lithium\util\Set::flatten($this->scaffold->object->data())]); ?>